# apps/studio/frontend/backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Junjo AI Studio",
    "description": "The Junjo AI Studio API: observability, execution evidence, and evaluation control.",
    "license": {
      "name": "Apache-2.0",
      "identifier": "Apache-2.0"
    },
    "version": "0.86.0"
  },
  "paths": {
...
    "/api/v1/evaluation/attempts/{attempt_id}": {
      "get": {
        "tags": [
          "evaluation"
        ],
        "summary": "Get one attempt with its run, dataset, and case.",
        "operationId": "get_evaluation_attempt",
        "parameters": [
          {
            "name": "attempt_id",
            "in": "path",
            "description": "The attempt's identifier.",
            "required": true,
            "schema": {
              "type": "string",
              "maxLength": 64,
              "minLength": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The attempt.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvaluationAttemptDetail"
                }
              }
            }
          },
          "401": {
            "description": "No valid credential.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "The token lacks the evaluation read scope.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Attempt not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "The request is not valid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "EvaluationControlToken": []
          }
        ]
      }
    },
...
    "/api/v1/evaluation/datasets": {
      "get": {
        "tags": [
          "evaluation"
        ],
        "summary": "List datasets, newest first.",
        "operationId": "list_evaluation_datasets",
        "parameters": [
          {
            "name": "application_key",
            "in": "query",
            "description": "List only this application's datasets.",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 128,
              "minLength": 1
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "description": "The `next_cursor` of the previous page.",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 1024,
              "minLength": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "The most datasets to return.",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 50,
              "maximum": 100,
              "minimum": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "One page of datasets.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvaluationDatasetList"
                }
              }
            }
          },
          "401": {
            "description": "No valid credential.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "The token lacks the evaluation read scope.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "The request is not valid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "EvaluationControlToken": []
          }
        ]
      },
      "post": {
        "tags": [
          "evaluation"
        ],
        "summary": "Create a dataset.",
        "description": "A repeated request returns the dataset it created.",
        "operationId": "create_evaluation_dataset",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EvaluationDatasetCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "The dataset.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvaluationDatasetRead"
                }
              }
            }
          },
          "401": {
            "description": "No valid credential.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "The token lacks the evaluation write scope.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "The dataset key exists with different content.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvaluationConflictResponse"
                }
              }
            }
          },
          "422": {
            "description": "The request is not valid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "EvaluationControlToken": []
          }
        ]
      }
    },
    "/api/v1/evaluation/datasets/{dataset_id}": {
      "get": {
        "tags": [
          "evaluation"
        ],
        "summary": "Get one dataset with its cases in order.",
        "operationId": "get_evaluation_dataset",
        "parameters": [
          {
            "name": "dataset_id",
            "in": "path",
            "description": "The dataset's identifier.",
            "required": true,
            "schema": {
              "type": "string",
              "maxLength": 64,
              "minLength": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The dataset and its cases.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvaluationDatasetDetail"
                }
              }
            }
          },
          "401": {
            "description": "No valid credential.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "The token lacks the evaluation read scope.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Dataset not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "The request is not valid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "EvaluationControlToken": []
          }
        ]
      }
    },
...
    "/api/v1/evaluation/runs": {
      "get": {
        "tags": [
          "evaluation"
        ],
        "summary": "List runs, newest first.",
        "description": "A run is listed when one case of its dataset matches every scope parameter\nsent.",
        "operationId": "list_evaluation_runs",
        "parameters": [
          {
            "name": "dataset_id",
            "in": "query",
            "description": "List only this dataset's runs.",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 64,
              "minLength": 1
            }
          },
          {
            "name": "target_kind",
            "in": "query",
            "description": "List only runs with a case of this target kind.",
            "required": false,
            "schema": {
              "type": "string",
              "description": "The kind of application target a case runs.",
              "enum": [
                "node",
                "workflow",
                "agent"
              ]
            }
          },
          {
            "name": "target_key",
            "in": "query",
            "description": "List only runs with a case of this target.",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 128,
              "minLength": 1
            }
          },
          {
            "name": "input_version",
            "in": "query",
            "description": "List only runs with a case of this input version.",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 2147483647,
              "minimum": 1
            }
          },
          {
            "name": "evaluation_name",
            "in": "query",
            "description": "List only runs with a case of this evaluation name.",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 256,
              "minLength": 1
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "description": "The `next_cursor` of the previous page.",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 1024,
              "minLength": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "The most runs to return.",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 50,
              "maximum": 100,
              "minimum": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "One page of runs.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvaluationRunList"
                }
              }
            }
          },
          "401": {
            "description": "No valid credential.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "The token lacks the evaluation read scope.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "The request is not valid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "EvaluationControlToken": []
          }
        ]
      },
      "post": {
        "tags": [
          "evaluation"
        ],
        "summary": "Start a run of a locked dataset.",
        "description": "A repeated request returns the run it started.",
        "operationId": "start_evaluation_run",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EvaluationRunStart"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "The run, with one attempt per case.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvaluationRunDetail"
                }
              }
            }
          },
          "401": {
            "description": "No valid credential.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "The token lacks the evaluation write scope.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Dataset not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "The request key exists with different content, or the dataset is not locked or has no case.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvaluationConflictResponse"
                }
              }
            }
          },
          "422": {
            "description": "The request is not valid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "EvaluationControlToken": []
          }
        ]
      }
    },
...
    "/api/v1/observability/traces/{trace_id}/spans": {
      "get": {
        "tags": [
          "observability"
        ],
        "summary": "Get every span of a trace, newest first.",
        "operationId": "get_trace_spans_api_v1_observability_traces__trace_id__spans_get",
        "parameters": [
          {
            "name": "trace_id",
            "in": "path",
            "description": "The trace's identifier.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The trace's spans, newest first.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                }
              }
            }
          },
          "401": {
            "description": "No signed-in user.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "description": "The body of every error response.",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "string",
            "description": "A stable machine-readable code.",
            "examples": [
              "not_found"
            ]
          },
          "message": {
            "type": "string",
            "description": "A human-readable explanation."
          }
        }
      },
      "EvaluationAttemptDetail": {
        "type": "object",
        "description": "One attempt with its run, dataset, and case.",
        "required": [
          "run",
          "dataset",
          "case",
          "attempt"
        ],
        "properties": {
          "run": {
            "$ref": "#/components/schemas/EvaluationRunRead"
          },
          "dataset": {
            "$ref": "#/components/schemas/EvaluationDatasetRead"
          },
          "case": {
            "$ref": "#/components/schemas/EvaluationCaseRead"
          },
          "attempt": {
            "$ref": "#/components/schemas/EvaluationAttemptRead"
          }
        },
        "additionalProperties": false
      },
...
      "EvaluationConflictResponse": {
        "type": "object",
        "description": "The body of a conflict response. `code` names the conflict. It describes\nwhat `ApiError::conflict` builds and is never constructed.",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "EvaluationControlToken": {
        "type": "http",
        "scheme": "bearer",
        "description": "A separately scoped Studio evaluation-control token. Studio ingestion API keys are not accepted."
      }
    }
  }
}